made a change regarding the data save intervals, so now the data save interval reduces to 0.01 seconds if the particle is very close to the asteroid and otherwise the data save interval value remains a large value.

+++ include/NAOS/particleAroundUniformlyRotatingEllipsoid.hpp
@@ -86,7 +86,7 @@ void executeSingleRegolithTrajectoryCalculation( const double alpha,
                                                  const std::vector< double > initialSunOrbitalElements,
                                                  const double sunMeanMotion,
                                                  SQLite::Statement &databaseQuery,
-                                                 const int dataSaveIntervals );
+                                                 const double dataSaveIntervals );
 
 //! convert body frame state vector to inertial frame for a uniformly rotating case
 /*!
